+++ packages/babel-preset/web.js
@@ -3,10 +3,10 @@ const nonStandardPlugins = require('./non-standard-plugins');
 module.exports = function shopifyWebPreset(_api, options = {}) {
   const {
     modules = 'commonjs',
-    corejs = 2,
+    corejs = 3,
     debug = false,
     browsers,
-    useBuiltIns = 'entry',
+    useBuiltIns = 'usage',
     typescript = false,
   } = options;